split out the single applicaiton classes from the main commotion_client code file
